Release 0.18.0.

(See also https://github.com/GoogleCloudPlatform/PerfKitBenchmarker/issues/369
which includes this change log with clickable GH-* links.)

* New features:
  * Support OpenStack as cloud provider (GH-305, GH-353, thanks @kivio and
    @mateusz-blaszkowski)
  * Support Rackspace as cloud provider (GH-336, thanks @meteorfox and @jrperritt)
  * Add support for ContainerizedVM using docker exec (GH-333, thanks @gablg1)
  * Windows guest VM support on Static VM (GH-350), Azure (GH-349, GH-374), AWS
    (GH-347), and GCE (GH-338)
  * Add NTttcp Windows networking benchmark (GH-348)

* Enhancements:
  * Support using proxies in VMs (GH-339, GH-337, thanks @kivio)
  * Enable optional migration on GCE (GH-343)
  * Execute long running commands via a remote agent (GH-310)
  * Add resource creation/deletion times to logs (GH-316)

* Bugfixes and maintenance updates:
  * Update PKB to work with Azure version 0.9.3 (GH-312)
  * Fix AWS CLI usage on Windows host (GH-313)
  * Auto-fetch AMI IDs for AWS images (GH-364)
  * Fix publisher missing info for default image and machine type (GH-357)
  * Fix 'no attribute pkb_thread_log_context' error for sub-thread logs (GH-322)

* Benchmark-specific changes:
  * aerospike: config/flag handling bugfixes (GH-367, GH-360, GH-354)
  * cassandra_ycsb: move num_vms prerequisite check
  * fio: add latency percentiles for results (GH-344)
  * hadoop_terasort: Fix bad SSH option (GH-328)
  * iperf: add lower bounds to arguments (GH-314)
  * iperf: add timeout to parallel benchmark runs to handle iperf hangs (GH-375)
  * netperf: Support confidence intervals, increase test length, report stddev
    (GH-317, GH-306)
  * ycsb: Drop unaggregatable results from samples (GH-324)

* Development and testing:
  * **Breaking Change** Automated testing now uses `tox` (GH-330)
  * Refactored hook scripts, including new opt-in pre-push hook (GH-363)
  * Use travis for CI testing (GH-340)
  * Speed up tests using timeouts (GH-299)

* Internals:
  * Move defaults from benchmark_spec to VM classes, move network instantiation
    out of benchmark spec (GH-342)
  * Add event hook support (GH-315)
  * Refactor VM classes (GH-321)

+++ perfkitbenchmarker/vm_util.py
@@ -15,22 +15,26 @@
 """Set of utility functions for working with virtual machines."""
 
 import contextlib
+import functools
 import logging
 import os
 import posixpath
 import random
 import re
 import socket
+import string
 import subprocess
 import tempfile
 import threading
 import time
 import traceback
+import uuid
 
 import jinja2
 
 from perfkitbenchmarker import data
 from perfkitbenchmarker import errors
+from perfkitbenchmarker import events
 from perfkitbenchmarker import flags
 from perfkitbenchmarker import log_util
 from perfkitbenchmarker import regex_util
@@ -57,6 +61,7 @@ FUZZ = .5
 MAX_RETRIES = -1
 
 WINDOWS = 'nt'
+PASSWORD_LENGTH = 15
 
 flags.DEFINE_integer('default_timeout', TIMEOUT, 'The default timeout for '
                      'retryable commands in seconds.')
@@ -70,6 +75,11 @@ flags.DEFINE_boolean('dstat', False,
 flags.DEFINE_integer('dstat_interval', None,
                      'dstat sample collection frequency, in seconds. Only '
                      'applicable when --dstat is specified.')
+flags.DEFINE_string('dstat_output', None,
+                    'Output directory for dstat output. '
+                    'Only applicable when --dstat is specified. '
+                    'Default: run temporary directory.')
+
 
 
 class IpAddressSubset(object):
@@ -236,17 +246,20 @@ def RunThreaded(target, thread_params, max_concurrent_threads=200):
             for i in range(0, 10)]
     RunThreaded(MyThreadedTargetMethod, args)
   """
-  if not thread_params:
-    raise ValueError('Param "thread_params" can\'t be empty')
-
   if not isinstance(thread_params, list):
     raise ValueError('Param "thread_params" must be a list')
+
+  if not thread_params:
+    # Nothing to do.
+    return
 
   if not isinstance(thread_params[0], tuple):
     thread_params = [((arg,), {}) for arg in thread_params]
   elif (not isinstance(thread_params[0][0], tuple) or
         not isinstance(thread_params[0][1], dict)):
     raise ValueError('If Param is a tuple, the tuple must be (tuple, dict)')
+  else:
+    thread_params = thread_params[:]
 
   threads = []
   exceptions = []
@@ -437,6 +450,7 @@ def IssueRetryableCommand(cmd, env=None):
   Args:
     cmd: A list of strings such as is given to the subprocess.Popen()
         constructor.
+    env: An alternate environment to pass to the Popen command.
 
   Returns:
     A tuple of stdout and stderr from running the provided command.
@@ -461,26 +475,6 @@ def ParseTimeCommandResult(command_result):
   time_in_seconds = 60 * float(time_data[0][0]) + float(time_data[0][1])
   return time_in_seconds
 
-
-def BurnCpu(vm, burn_cpu_threads=None, burn_cpu_seconds=None):
-  """Burns vm cpu for some amount of time and dirty cache.
-
-  Args:
-    vm: The target vm.
-    burn_cpu_threads: Number of threads to burn cpu.
-    burn_cpu_seconds: Amount of time in seconds to burn cpu.
-  """
-  burn_cpu_threads = burn_cpu_threads or FLAGS.burn_cpu_threads
-  burn_cpu_seconds = burn_cpu_seconds or FLAGS.burn_cpu_seconds
-  if burn_cpu_seconds:
-    vm.Install('sysbench')
-    end_time = time.time() + burn_cpu_seconds
-    vm.RemoteCommand(
-        'nohup sysbench --num-threads=%s --test=cpu --cpu-max-prime=10000000 '
-        'run 1> /dev/null 2> /dev/null &' % burn_cpu_threads)
-    if time.time() < end_time:
-      time.sleep(end_time - time.time())
-    vm.RemoteCommand('pkill -9 sysbench')
 
 
 def ShouldRunOnExternalIpAddress():
@@ -528,7 +522,7 @@ def GetLastRunUri():
 
 
 @contextlib.contextmanager
-def NamedTemporaryFile(prefix='tmp', suffix='', dir=None):
+def NamedTemporaryFile(prefix='tmp', suffix='', dir=None, delete=True):
   """Behaves like tempfile.NamedTemporaryFile.
 
   The existing tempfile.NamedTemporaryFile has the annoying property on
@@ -545,7 +539,8 @@ def NamedTemporaryFile(prefix='tmp', suffix='', dir=None):
   finally:
     if not f.closed:
       f.close()
-    os.unlink(f.name)
+    if delete:
+      os.unlink(f.name)
 
 
 def GenerateSSHConfig(vms):
@@ -590,28 +585,31 @@ def ExecutableOnPath(executable_name):
   return True
 
 
-@contextlib.contextmanager
-def RunDStatIfConfigured(vms, suffix='-dstat'):
-  """Installs and runs dstat on 'vms' if FLAGS.dstat is set.
+class _DStatCollector(object):
+  """dstat collector.
 
-  On exiting the context manager, the results are copied to the run temp dir.
-
-  Args:
-    vms: List of virtual machines.
-    suffix: str. Suffix to add to each dstat result file.
-
-  Yields:
-    None
+  Installs and runs dstat on a collection of VMs.
   """
-  if not FLAGS.dstat:
-    yield
-    return
 
-  lock = threading.Lock()
-  pids = {}
-  file_names = {}
+  def __init__(self, interval=None, output_directory=None):
+    """Runs dstat on 'vms'.
 
-  def StartDStat(vm):
+    Start dstat collection via `Start`. Stop via `Stop`.
+
+    Args:
+      interval: Optional int. Interval in seconds in which to collect samples.
+    """
+    self.interval = interval
+    self.output_directory = output_directory or GetTempDir()
+    self._lock = threading.Lock()
+    self._pids = {}
+    self._file_names = {}
+
+    if not os.path.isdir(self.output_directory):
+      raise IOError('dstat output directory does not exist: {0}'.format(
+          self.output_directory))
+
+  def _StartOnVm(self, vm, suffix='-dstat'):
     vm.Install('dstat')
 
     num_cpus = vm.num_cpus
@@ -631,25 +629,75 @@ def RunDStatIfConfigured(vms, suffix='-dstat'):
                max_cpu=num_cpus - 1,
                block_devices=','.join(block_devices),
                output=dstat_file,
-               dstat_interval=FLAGS.dstat_interval or '')
+               dstat_interval=self.interval or '')
     stdout, _ = vm.RemoteCommand(cmd)
-    with lock:
-      pids[vm.name] = stdout.strip()
-      file_names[vm.name] = dstat_file
+    with self._lock:
+      self._pids[vm.name] = stdout.strip()
+      self._file_names[vm.name] = dstat_file
 
-  def StopDStat(vm):
-    if vm.name not in pids:
+  def _StopOnVm(self, vm):
+    """Stop dstat on 'vm', copy the results to the run temporary directory."""
+    if vm.name not in self._pids:
       logging.warn('No dstat PID for %s', vm.name)
       return
-    cmd = 'kill {0} || true'.format(pids[vm.name])
+    else:
+      with self._lock:
+        pid = self._pids.pop(vm.name)
+        file_name = self._file_names.pop(vm.name)
+    cmd = 'kill {0} || true'.format(pid)
     vm.RemoteCommand(cmd)
     try:
-      vm.PullFile(GetTempDir(), file_names[vm.name])
+      vm.PullFile(self.output_directory, file_name)
     except:
-      logging.exception('Failed fetching dstat result.')
+      logging.exception('Failed fetching dstat result from %s.', vm.name)
 
-  RunThreaded(StartDStat, vms)
-  try:
-    yield
-  finally:
-    RunThreaded(StopDStat, vms)
+  def Start(self, sender, benchmark_spec):
+    """Install and start dstat on all VMs in 'benchmark_spec'."""
+    suffix = '-{0}-{1}-dstat'.format(benchmark_spec.benchmark_name,
+                                     str(uuid.uuid4())[:8])
+    start_on_vm = functools.partial(self._StartOnVm, suffix=suffix)
+    RunThreaded(start_on_vm, benchmark_spec.vms)
+
+  def Stop(self, sender, benchmark_spec):
+    """Stop dstat on all VMs in 'benchmark_spec', fetch results."""
+    RunThreaded(self._StopOnVm, benchmark_spec.vms)
+
+
+@events.initialization_complete.connect
+def _RegisterDStatCollector(sender, parsed_flags):
+  """Registers the dstat collector if FLAGS.dstat is set."""
+  if not parsed_flags.dstat:
+    return
+
+  output_directory = (parsed_flags.dstat_output
+                      if parsed_flags['dstat_output'].present
+                      else GetTempDir())
+
+  logging.debug('Registering dstat collector with interval %s, output to %s.',
+                parsed_flags.dstat_interval, output_directory)
+
+  if not os.path.isdir(output_directory):
+    os.makedirs(output_directory)
+  collector = _DStatCollector(interval=parsed_flags.dstat_interval,
+                              output_directory=output_directory)
+  events.before_phase.connect(collector.Start, events.RUN_PHASE, weak=False)
+  events.after_phase.connect(collector.Stop, events.RUN_PHASE, weak=False)
+
+
+def GenerateRandomWindowsPassword(password_length=PASSWORD_LENGTH):
+  """Generates a password that meets Windows complexity requirements."""
+  # The special characters have to be recognized by the Azure CLI as
+  # special characters. This greatly limits the set of characters
+  # that we can safely use. See
+  # https://github.com/Azure/azure-xplat-cli/blob/master/lib/commands/arm/vm/vmOsProfile._js#L145
+  special_chars = '*!@#$%^+='
+  password = [
+      random.choice(string.ascii_letters + string.digits + special_chars)
+      for _ in range(password_length - 4)]
+  # Ensure that the password contains at least one of each 4 required
+  # character types.
+  password.append(random.choice(string.ascii_lowercase))
+  password.append(random.choice(string.ascii_uppercase))
+  password.append(random.choice(string.digits))
+  password.append(random.choice(special_chars))
+  return ''.join(password)
